requirements.txt
annotated-types==0.7.0
anyio==4.10.0
APScheduler==3.11.0
blockbuster==1.5.25
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
cryptography==44.0.3
distro==1.9.0
dnspython==2.8.0
fastapi==0.116.2
forbiddenfruit==0.1.4
genai-prices==0.0.27
google-auth==2.40.3
google-genai==1.38.0
greenlet==3.2.4
griffe==1.14.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
jiter==0.11.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema_rs==0.29.1
langchain==0.3.27
langchain-core==0.3.76
langchain-mongodb==0.7.0
langchain-text-splitters==0.3.11
langgraph==0.6.7
langgraph-api==0.4.20
langgraph-checkpoint==2.1.1
langgraph-checkpoint-mongodb==0.2.0
langgraph-cli==0.4.2
langgraph-prebuilt==0.6.4
langgraph-runtime-inmem==0.12.0
langgraph-sdk==0.2.7
langsmith==0.4.28
lark==1.2.2
logfire-api==4.7.0
numpy==2.3.3
openai==1.107.3
opentelemetry-api==1.37.0
orjson==3.11.3
ormsgpack==1.10.0
packaging==25.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybreaker==1.4.0
pycparser==2.23
pydantic==2.11.9
pydantic-ai-slim==1.0.8
pydantic-graph==1.0.8
pydantic_core==2.33.2
PyJWT==2.10.1
pymongo==4.14.1
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
setuptools==78.1.1
sniffio==1.3.1
SQLAlchemy==2.0.43
sse-starlette==2.1.3
starlette==0.48.0
structlog==25.4.0
tenacity==9.1.2
tqdm==4.67.1
truststore==0.10.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
wheel==0.45.1
xxhash==3.5.0
zipp==3.23.0
zstandard==0.25.0
